add : fix type

+++ tmdb-web/src/main/java/com/ktdsuniversity/edu/movie/vo/MovieVO.java
@@ -15,8 +15,8 @@ public class MovieVO {
 	private String originalTitle;
 	private String state;
 	private String language;
-	private int budget;
-	private int profit;
+	private long budget;
+	private long profit;
 	
 	public String getMovieId() {
 		return this.movieId;
@@ -90,16 +90,16 @@ public class MovieVO {
 	public void setLanguage(String language) {
 		this.language = language;
 	}
-	public int getBudget() {
+	public long getBudget() {
 		return this.budget;
 	}
-	public void setBudget(int budget) {
+	public void setBudget(long budget) {
 		this.budget = budget;
 	}
-	public int getProfit() {
+	public long getProfit() {
 		return this.profit;
 	}
-	public void setProfit(int profit) {
+	public void setProfit(long profit) {
 		this.profit = profit;
 	}
